backend/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pyyaml==6.0.1
httpx==0.25.2
notion-client==2.2.1

# Audio Processing
pvporcupine==3.0.0
pyaudio==0.2.14
openai==1.51.0
openai-whisper==20231117
sounddevice==0.4.6
webrtcvad==2.0.10
numpy==1.24.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Optional Database Dependencies (not needed with Notion API)
# Uncomment these if you want to add database support later:
# sqlalchemy==2.0.23
# alembic==1.13.0
# redis==5.0.1
# psycopg2-binary==2.9.9
# python-jose[cryptography]==3.3.0
# bcrypt==4.1.2